generate metadata upon initialisation
