+++ changelog.md
@@ -5,6 +5,7 @@ v0.0.11:
   - adds `dfschema` and `pandas` version in metadata upon generation (Later will worn if Schema is initialized from json, generated by later version)
 - Renamed `na_limit` to `na_pct_below` to make it unambiguous (with backward support)
 - Added `optional=True` flag for columns. If true, does not raise exception if column is not present
+- added `dfschema update --input {existing_schema} --output {output_schema}` command to upgrade schemas
 
 v0.0.10:
 - relaxed Pydantic requirement to `>=1.9`
@@ -37,7 +38,7 @@ v0.0.6:
     - added pre-commit install to the repo
     - Some benchmarking
     - renamed `dfs.validate_df` to `dfs.validate`
-    
+
 v0.0.5: fix column dtype generation/validation bug
 
 ## Pre-Publication
